+++ Whim.sln
@@ -37,8 +37,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Whim.CommandPalette", "src\Whim.CommandPalette\Whim.CommandPalette.csproj", "{05AB1322-761E-4C82-BF41-9613CA563C5A}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Whim.CommandPalette.Tests", "src\Whim.CommandPalette.Tests\Whim.CommandPalette.Tests.csproj", "{867E9712-E9FC-4C27-8014-1FD660236BBB}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Whim.TreeLayout.Bar", "src\Whim.TreeLayout.Bar\Whim.TreeLayout.Bar.csproj", "{4F66FB22-C4B3-494F-BEE0-9F878349621E}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Whim.FloatingLayout.Tests", "src\Whim.FloatingLayout.Tests\Whim.FloatingLayout.Tests.csproj", "{2E407C3C-FCE9-4851-9091-61C6CC687AFD}"
@@ -47,7 +45,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Whim.Gaps.Tests", "src\Whim
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Whim.TreeLayout.CommandPalette", "src\Whim.TreeLayout.CommandPalette\Whim.TreeLayout.CommandPalette.csproj", "{8CD01B7A-78BD-48BD-B274-B2FEE48CC8F4}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Whim.CommandPalette.WinUITests", "src\Whim.CommandPalette.WinUITests\Whim.CommandPalette.WinUITests.csproj", "{3A63AEA5-DAFF-4712-B7F7-0E3D5320636A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Whim.CommandPalette.Tests", "src\Whim.CommandPalette.Tests\Whim.CommandPalette.Tests.csproj", "{3A63AEA5-DAFF-4712-B7F7-0E3D5320636A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -225,22 +223,6 @@ Global
 		{05AB1322-761E-4C82-BF41-9613CA563C5A}.Release|x64.Build.0 = Release|x64
 		{05AB1322-761E-4C82-BF41-9613CA563C5A}.Release|x86.ActiveCfg = Release|Any CPU
 		{05AB1322-761E-4C82-BF41-9613CA563C5A}.Release|x86.Build.0 = Release|Any CPU
-		{867E9712-E9FC-4C27-8014-1FD660236BBB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{867E9712-E9FC-4C27-8014-1FD660236BBB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{867E9712-E9FC-4C27-8014-1FD660236BBB}.Debug|arm64.ActiveCfg = Debug|arm64
-		{867E9712-E9FC-4C27-8014-1FD660236BBB}.Debug|arm64.Build.0 = Debug|arm64
-		{867E9712-E9FC-4C27-8014-1FD660236BBB}.Debug|x64.ActiveCfg = Debug|x64
-		{867E9712-E9FC-4C27-8014-1FD660236BBB}.Debug|x64.Build.0 = Debug|x64
-		{867E9712-E9FC-4C27-8014-1FD660236BBB}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{867E9712-E9FC-4C27-8014-1FD660236BBB}.Debug|x86.Build.0 = Debug|Any CPU
-		{867E9712-E9FC-4C27-8014-1FD660236BBB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{867E9712-E9FC-4C27-8014-1FD660236BBB}.Release|Any CPU.Build.0 = Release|Any CPU
-		{867E9712-E9FC-4C27-8014-1FD660236BBB}.Release|arm64.ActiveCfg = Release|arm64
-		{867E9712-E9FC-4C27-8014-1FD660236BBB}.Release|arm64.Build.0 = Release|arm64
-		{867E9712-E9FC-4C27-8014-1FD660236BBB}.Release|x64.ActiveCfg = Release|x64
-		{867E9712-E9FC-4C27-8014-1FD660236BBB}.Release|x64.Build.0 = Release|x64
-		{867E9712-E9FC-4C27-8014-1FD660236BBB}.Release|x86.ActiveCfg = Release|Any CPU
-		{867E9712-E9FC-4C27-8014-1FD660236BBB}.Release|x86.Build.0 = Release|Any CPU
 		{4F66FB22-C4B3-494F-BEE0-9F878349621E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{4F66FB22-C4B3-494F-BEE0-9F878349621E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4F66FB22-C4B3-494F-BEE0-9F878349621E}.Debug|arm64.ActiveCfg = Debug|arm64
