+++ src/Whim.CommandPalette.Tests/App.xaml
@@ -1,0 +1,5 @@
+﻿<Application
+	x:Class="Whim.CommandPalette.Tests.App"
+	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+	xmlns:local="using:Whim.CommandPalette.Tests" />
