+++ src/Whim.CommandPalette.Tests/CommandPaletteCommandsTests.cs
