+++ src/Whim.CommandPalette.Tests/CommandPalettePluginTests.cs
