--- src/Whim.CommandPalette.Tests/Directory.Build.props
@@ -1,5 +1,0 @@
-<Project>
-	<PropertyGroup>
-		<GenerateDocumentationFile>false</GenerateDocumentationFile>
-	</PropertyGroup>
-</Project>
