+++ src/Whim.CommandPalette.Tests/MainWindow.xaml
@@ -1,0 +1,11 @@
+﻿<Window
+	x:Class="Whim.CommandPalette.Tests.MainWindow"
+	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+	xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+	xmlns:local="using:Whim.CommandPalette.Tests"
+	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+	mc:Ignorable="d">
+
+	<Frame />
+</Window>
