+++ src/Whim.CommandPalette.Tests/MainWindow.xaml.cs
@@ -1,0 +1,11 @@
+using Microsoft.UI.Xaml;
+
+namespace Whim.CommandPalette.Tests;
+
+public sealed partial class MainWindow : Window
+{
+	public MainWindow()
+	{
+		InitializeComponent();
+	}
+}
