+++ src/Whim.CommandPalette.Tests/MatchTests.cs
