+++ src/Whim.CommandPalette.Tests/MostRecentlyUsedMatcherTests.cs
