+++ src/Whim.CommandPalette.Tests/Variants/Menu/MenuRowStub.cs
@@ -1,0 +1,16 @@
+namespace Whim.CommandPalette.Tests;
+
+internal class MenuRowStub : IVariantRowView<CommandItem, MenuVariantRowViewModel>
+{
+	public bool IsUpdated { get; private set; }
+
+	public required MenuVariantRowViewModel ViewModel { get; init; }
+
+	public void Initialize() { }
+
+	public void Update(MatcherResult<CommandItem> matcherResult)
+	{
+		ViewModel.Update(matcherResult);
+		IsUpdated = true;
+	}
+}
