+++ src/Whim.CommandPalette.Tests/app.manifest
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="utf-8"?>
+<assembly manifestVersion="1.0" xmlns="urn:schemas-microsoft-com:asm.v1">
+  <assemblyIdentity version="1.0.0.0" name="Whim.CommandPalette.Tests.app" />
+
+  <compatibility xmlns="urn:schemas-microsoft-com:compatibility.v1">
+    <application>
+      <!--The
+      ID below informs the system that this application is compatible with OS features first
+      introduced in Windows 8.
+      For more info see
+      https://docs.microsoft.com/windows/win32/sysinfo/targeting-your-application-at-windows-8-1
+
+      It is also necessary to support features in unpackaged applications, for example the custom
+      titlebar implementation.-->
+      <supportedOS Id="{4a2f28e3-53b9-4441-ba9c-d69d4a4a6e38}" />
+    </application>
+  </compatibility>
+
+  <application xmlns="urn:schemas-microsoft-com:asm.v3">
+    <windowsSettings>
+      <!-- The combination of below two tags have the following effect:
+           1) Per-Monitor for >= Windows 10 Anniversary Update
+           2) System < Windows 10 Anniversary Update
+      -->
+      <dpiAware xmlns="http://schemas.microsoft.com/SMI/2005/WindowsSettings">true/PM</dpiAware>
+      <dpiAwareness xmlns="http://schemas.microsoft.com/SMI/2016/WindowsSettings">PerMonitorV2,
+        PerMonitor</dpiAwareness>
+    </windowsSettings>
+  </application>
+</assembly>
