--- src/Whim.CommandPalette.WinUITests/App.xaml
@@ -1,5 +1,0 @@
-﻿<Application
-	x:Class="Whim.CommandPalette.WinUITests.App"
-	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-	xmlns:local="using:Whim.CommandPalette.WinUITests" />
