--- src/Whim.CommandPalette.WinUITests/App.xaml.cs
@@ -1,41 +1,0 @@
-﻿using Microsoft.UI.Xaml;
-using Microsoft.VisualStudio.TestTools.UnitTesting.AppContainer;
-using System;
-
-namespace Whim.CommandPalette.WinUITests;
-
-/// <summary>
-/// Provides application-specific behavior to supplement the default Application class.
-/// </summary>
-public partial class App : Application
-{
-	/// <summary>
-	/// Initializes the singleton application object.  This is the first line of authored code
-	/// executed, and as such is the logical equivalent of main() or WinMain().
-	/// </summary>
-	public App()
-	{
-		this.InitializeComponent();
-	}
-
-	/// <summary>
-	/// Invoked when the application is launched.
-	/// </summary>
-	/// <param name="args">Details about the launch request and process.</param>
-	protected override void OnLaunched(LaunchActivatedEventArgs args)
-	{
-		Microsoft.VisualStudio.TestPlatform.TestExecutor.UnitTestClient.CreateDefaultUI();
-
-		m_window = new MainWindow();
-
-		// Ensure the current window is active
-		m_window.Activate();
-
-		UITestMethodAttribute.DispatcherQueue = m_window.DispatcherQueue;
-
-		// Replace back with e.Arguments when https://github.com/microsoft/microsoft-ui-xaml/issues/3368 is fixed
-		Microsoft.VisualStudio.TestPlatform.TestExecutor.UnitTestClient.Run(Environment.CommandLine);
-	}
-
-	private Window? m_window;
-}
