--- src/Whim.CommandPalette.WinUITests/MainWindow.xaml.cs
@@ -1,11 +1,0 @@
-using Microsoft.UI.Xaml;
-
-namespace Whim.CommandPalette.WinUITests;
-
-public sealed partial class MainWindow : Window
-{
-	public MainWindow()
-	{
-		InitializeComponent();
-	}
-}
