--- src/Whim.CommandPalette.WinUITests/Properties/launchSettings.json
@@ -1,10 +1,0 @@
-{
-  "profiles": {
-    "Whim.CommandPalette.WinUITests (Package)": {
-      "commandName": "MsixPackage"
-    },
-    "Whim.CommandPalette.WinUITests (Unpackaged)": {
-      "commandName": "Project"
-    }
-  }
-}
